fix: share exact-role worker cli resolver
